requirements.txt
beautifulsoup4>=4.12.0
curl-cffi>=0.7.0
lxml>=5.0.0
Pillow>=10.0.0
